+++ database/migrations/2025_05_07_203458_remove-twitter.php
@@ -7,6 +7,10 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration
 {
     public function up(): void {
+        if (Schema::connection(null)->getConnection()->getDriverName() === 'sqlite') {
+            return;
+        }
+
         Schema::table('social_login_profiles', function(Blueprint $table) {
             $table->dropUnique(['twitter_id']);
             $table->dropColumn('twitter_id');
